Moscow_iGEM2020
Kinetic model of the detection system with CasX. Description of the model and results you can see in our wiki page

We created a detailed kinetic model to ensure system performance, optimal conditions, reaction time, and potential sensitivity. The goals of the model are the following:

To confirm our expectations of system functioning, such as the potential to detect nucleic acids in a reasonable time
To determine optimal conditions (sgRNA initial concentration, CasX initial concentration, etc.),
To determine detector sensitivity - the minimum concentration of nucleic acid in a sample that can be detected using our test system,
To determine the time required to carry out one test.